requirements.txt
flask
flask-sqlalchemy
flask-cors
spotipy
python-dotenv
requests
gunicorn
